+++ frontend/src/redux/rootSaga.ts
@@ -1,8 +1,10 @@
 import { all } from "redux-saga/effects";
 import { authWatcher } from "./ducks/auth/saga";
+import { profileWatcher } from "./ducks/profile/saga";
 
 export function* rootSaga() {
     yield  all([
-        authWatcher()
+        authWatcher(),
+        profileWatcher()
     ])
 }